+++ airbyte-cdk/python/airbyte_cdk/sources/file_based/config/csv_format.py
@@ -147,11 +147,16 @@ class CsvFormat(BaseModel):
         description="How to infer the types of the columns. If none, inference default to strings.",
         airbyte_hidden=True,
     )
+    ignore_errors_on_fields_mismatch: bool = Field(
+        title="Ignore errors on field mismatch",
+        default=False,
+        description="Whether to ignore errors that occur when the number of fields in the CSV does not match the number of columns in the schema.",
+    )
 
     @validator("delimiter")
     def validate_delimiter(cls, v: str) -> str:
         if v == r"\t":
-            return v
+            v = "\t"
         if len(v) != 1:
             raise ValueError("delimiter should only be one character")
         if v in {"\r", "\n"}: