+++ airbyte-ci/connectors/connectors_qa/src/connectors_qa/checks/metadata.py
@@ -3,7 +3,8 @@
 
 import os
 
-from connector_ops.utils import Connector  # type: ignore
+import toml
+from connector_ops.utils import Connector, ConnectorLanguage  # type: ignore
 from connectors_qa import consts
 from connectors_qa.models import Check, CheckCategory, CheckResult
 from metadata_service.validators.metadata_validator import PRE_UPLOAD_VALIDATORS, ValidatorOptions, validate_and_load  # type: ignore
@@ -56,7 +57,25 @@ class CheckConnectorLanguageTag(MetadataCheck):
     PYTHON_LANGUAGE_TAG = "language:python"
     JAVA_LANGUAGE_TAG = "language:java"
 
+    def get_expected_language_tag(self, connector: Connector) -> str:
+        if (connector.code_directory / consts.SETUP_PY_FILE_NAME).exists() or (
+            connector.code_directory / consts.PYPROJECT_FILE_NAME
+        ).exists():
+            return self.PYTHON_LANGUAGE_TAG
+        elif (connector.code_directory / consts.GRADLE_FILE_NAME).exists():
+            return self.JAVA_LANGUAGE_TAG
+        else:
+            raise ValueError("Could not infer the language tag from the connector directory")
+
     def _run(self, connector: Connector) -> CheckResult:
+        try:
+            expected_language_tag = self.get_expected_language_tag(connector)
+        except ValueError:
+            return self.fail(
+                connector=connector,
+                message="Could not infer the language tag from the connector directory",
+            )
+
         current_language_tags = [t for t in connector.metadata.get("tags", []) if t.startswith("language:")]
         if not current_language_tags:
             return self.fail(
@@ -69,31 +88,69 @@ class CheckConnectorLanguageTag(MetadataCheck):
                 message=f"Multiple language tags found in the metadata file: {current_language_tags}",
             )
         current_language_tag = current_language_tags[0]
-
-        if (connector.code_directory / consts.SETUP_PY_FILE_NAME).exists() or (
-            connector.code_directory / consts.PYPROJECT_FILE_NAME
-        ).exists():
-            expected_language = self.PYTHON_LANGUAGE_TAG
-        elif (connector.code_directory / consts.GRADLE_FILE_NAME).exists():
-            expected_language = self.JAVA_LANGUAGE_TAG
-        else:
+        if current_language_tag != expected_language_tag:
             return self.fail(
                 connector=connector,
-                message="Could not infer the language tag from the connector directory",
-            )
-        if current_language_tag != expected_language:
-            return self.fail(
-                connector=connector,
-                message=f"Expected language tag '{expected_language}' in the {consts.METADATA_FILE_NAME} file, but found '{current_language_tag}'",
+                message=f"Expected language tag '{expected_language_tag}' in the {consts.METADATA_FILE_NAME} file, but found '{current_language_tag}'",
             )
         return self.pass_(
             connector=connector,
-            message=f"Language tag {expected_language} is present in the metadata file",
+            message=f"Language tag {expected_language_tag} is present in the metadata file",
+        )
+
+
+class CheckConnectorCDKTag(MetadataCheck):
+    name = "Python connectors must have a CDK tag in metadata"
+    description = f"Python connectors must have a CDK tag in their metadata. It must be set in the `tags` field in {consts.METADATA_FILE_NAME}. The values can be `cdk:low-code`, `cdk:python`, or `cdk:file`."
+    applies_to_connector_languages = [ConnectorLanguage.PYTHON, ConnectorLanguage.LOW_CODE]
+
+    class CDKTag:
+        LOW_CODE = "cdk:low-code"
+        PYTHON = "cdk:python"
+        FILE = "cdk:python-file-based"
+
+    def get_expected_cdk_tag(self, connector: Connector) -> str:
+        manifest_file = connector.code_directory / connector.technical_name.replace("-", "_") / consts.LOW_CODE_MANIFEST_FILE_NAME
+        pyproject_file = connector.code_directory / consts.PYPROJECT_FILE_NAME
+        setup_py_file = connector.code_directory / consts.SETUP_PY_FILE_NAME
+        if manifest_file.exists():
+            return self.CDKTag.LOW_CODE
+        if pyproject_file.exists():
+            pyproject = toml.load((connector.code_directory / consts.PYPROJECT_FILE_NAME))
+            cdk_deps = pyproject["tool"]["poetry"]["dependencies"].get("airbyte-cdk", None)
+            if cdk_deps and isinstance(cdk_deps, dict) and "file-based" in cdk_deps.get("extras", []):
+                return self.CDKTag.FILE
+        if setup_py_file.exists():
+            if "airbyte-cdk[file-based]" in (connector.code_directory / consts.SETUP_PY_FILE_NAME).read_text():
+                return self.CDKTag.FILE
+        return self.CDKTag.PYTHON
+
+    def _run(self, connector: Connector) -> CheckResult:
+        current_cdk_tags = [t for t in connector.metadata.get("tags", []) if t.startswith("cdk:")]
+        expected_cdk_tag = self.get_expected_cdk_tag(connector)
+        if not current_cdk_tags:
+            return self.fail(
+                connector=connector,
+                message="CDK tag is missing in the metadata file",
+            )
+        if len(current_cdk_tags) > 1:
+            return self.fail(
+                connector=connector,
+                message=f"Multiple CDK tags found in the metadata file: {current_cdk_tags}",
+            )
+        if current_cdk_tags[0] != expected_cdk_tag:
+            return self.fail(
+                connector=connector,
+                message=f"Expected CDK tag '{self.get_expected_cdk_tag(connector)}' in the {consts.METADATA_FILE_NAME} file, but found '{current_cdk_tags[0]}'",
+            )
+        return self.pass_(
+            connector=connector,
+            message=f"CDK tag {self.get_expected_cdk_tag(connector)} is present in the metadata file",
         )
 
 
 ENABLED_CHECKS = [
     ValidateMetadata(),
-    # Disabled until metadata are globally cleaned up
-    # CheckConnectorLanguageTag()
+    CheckConnectorLanguageTag(),
+    CheckConnectorCDKTag(),
 ]