+++ airbyte-integrations/connectors/source-bing-ads/unit_tests/integrations/test_goals_and_funnels_report.py
@@ -9,8 +9,8 @@ class TestGoalsAndFunnelsReportDailyStream(TestSuiteReportStream):
     records_number = 8
     state_file = "goals_and_funnels_report_daily_state"
     incremental_report_file = "goals_and_funnels_report_daily_incremental"
-    first_read_state = {"goals_and_funnels_report_daily": {"180535609": {"TimePeriod": "2023-12-17"}}}
-    second_read_state = {"goals_and_funnels_report_daily": {"180535609": {"TimePeriod": "2023-12-24"}}}
+    first_read_state = {"180535609": {"TimePeriod": "2023-12-17"}}
+    second_read_state = {"180535609": {"TimePeriod": "2023-12-24"}}
 
 
 class TestGoalsAndFunnelsReportHourlyStream(TestSuiteReportStream):
@@ -19,8 +19,8 @@ class TestGoalsAndFunnelsReportHourlyStream(TestSuiteReportStream):
     records_number = 24
     state_file = "goals_and_funnels_report_hourly_state"
     incremental_report_file = "goals_and_funnels_report_hourly_incremental"
-    first_read_state = {"goals_and_funnels_report_hourly": {"180535609": {"TimePeriod": "2023-11-12T00:00:00+00:00"}}}
-    second_read_state = {"goals_and_funnels_report_hourly": {"180535609": {"TimePeriod": "2023-11-13T00:00:00+00:00"}}}
+    first_read_state = {"180535609": {"TimePeriod": "2023-11-12T00:00:00+00:00"}}
+    second_read_state = {"180535609": {"TimePeriod": "2023-11-13T00:00:00+00:00"}}
 
 
 class TestGoalsAndFunnelsReportWeeklyStream(TestSuiteReportStream):
@@ -30,8 +30,8 @@ class TestGoalsAndFunnelsReportWeeklyStream(TestSuiteReportStream):
     second_read_records_number = 5
     state_file = "goals_and_funnels_report_weekly_state"
     incremental_report_file = "goals_and_funnels_report_weekly_incremental"
-    first_read_state = {"goals_and_funnels_report_weekly": {"180535609": {"TimePeriod": "2023-12-25"}}}
-    second_read_state = {"goals_and_funnels_report_weekly": {"180535609": {"TimePeriod": "2024-01-29"}}}
+    first_read_state = {"180535609": {"TimePeriod": "2023-12-25"}}
+    second_read_state = {"180535609": {"TimePeriod": "2024-01-29"}}
 
 
 class TestGoalsAndFunnelsReportMonthlyStream(TestSuiteReportStream):
@@ -40,5 +40,5 @@ class TestGoalsAndFunnelsReportMonthlyStream(TestSuiteReportStream):
     records_number = 6
     state_file = "goals_and_funnels_report_monthly_state"
     incremental_report_file = "goals_and_funnels_report_monthly_incremental"
-    first_read_state = {"goals_and_funnels_report_monthly": {"180535609": {"TimePeriod": "2023-09-01"}}}
-    second_read_state = {"goals_and_funnels_report_monthly": {"180535609": {"TimePeriod": "2024-03-01"}}}
+    first_read_state = {"180535609": {"TimePeriod": "2023-09-01"}}
+    second_read_state = {"180535609": {"TimePeriod": "2024-03-01"}}