+++ airbyte-cdk/python/CHANGELOG.md
@@ -1,5 +1,44 @@
 # Changelog
 
+## 0.73.0
+set minimum python version to 3.9
+
+## 0.72.2
+Connector Builder: have schema fields be nullable by default except from PK and cursor field
+
+## 0.72.1
+low code: add refresh_token_error handler to DeclarativeOauth2Authenticator
+
+## 0.72.0
+low-code: Allow defining custom schema loaders
+
+## 0.71.0
+Declarative datetime-based cursors now only derive state values from records that were read
+
+## 0.70.2
+low-code: remove superfluous sleep
+
+## 0.70.1
+File-based CDK: Fix tab delimiter configuration in CSV file type
+
+## 0.70.0
+testing
+
+## 0.69.2
+low-code: improve error message when a custom component cannot be found
+
+## 0.69.1
+Update mock server test entrypoint wrapper to use per-stream state
+
+## 0.69.0
+Include recordCount in stream state messages and final state message for full refresh syncs
+
+## 0.68.4
+low-code: update cartesian stream slice to emit typed StreamSlice
+
+## 0.68.3
+Low-code: adding a default value if a stream slice is None during read_records
+
 ## 0.68.2
 low-code: remove parent cursor compoent from incremental substreams' state message